// FrontEnd/src/app/models/register.model.ts
export class register {
    username: string;
    Usesurname: string;
    password: string;
    repassword: string;
    Usename: string;
    idcard: number;
    phoneNum: string;
    email: any;
    address: string;
}

export interface registerRequire {
    username: string;
    Usesurname: string;
    password: string;
    repassword: string;
    Usename: string;
    idcard: number;
    phoneNum: number;
    email: any;
    address: string;
}